Remove all remaining mod.rs modules
